E7 : got lamp to make noise and light up.
next: going to introduce an NPC to click and make noise... then also going to have the user make the lamp explode.
Also need to lower buzzing sound! (and find a way to have each lamp buzz, and buzz when the player is near instead of all the time. and stop buzzing when they are turned off)