P2 : trying to create classes to store my npcs and items
so that they can be called at the right moments in different states. For my program I will probably have them all for the simulation state only, but may need to work with states in funny ways.

also need to find out how to manage the items in the game versus the items in player inventory