P2 : It seems I may have succeeded in getting item inside player inventory!
need to tweak the code, stamp out potential bug holes...
I need to find a way to have the items exist all the time and also be manipulated by a temporary item to be taken out of the map, or in and out of player/npc inventories.
player inventory holds the whole object (so that it can be recognized by the NPCs object label). The npc's inventory is only an object label which coincides or not with the object in player's inventory. When player has object NPC desires, (which the program knows through the NPC object label), the item is recalled in temporary item variable, found in the inventory (search for inventory[i].name === `LABEL`;) and then spliced!
After a trade, player receives the object label, and the correct object is retrieved in the gameItemsList array with the new item label, and then this new item class object is pushed into the player inventory.